requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
Werkzeug==1.0.0
tabulate